feat: unify chain identity to u64 network magic from config
Replace the u32 chain_id selector with the u64 network magic sourced
from psy-genesis/config.json, generated at build time by psy_core/build.rs
with PSY_CONFIG_PATH override. Supported selectors: LocalDevnet,
PsyPublicTestnet, PsyMainnet; all others fail closed.

- psy_core: generated PSY_CHAIN_ID_* constants; removed duplicate
  parth_core constants and migrated the sole consumer
- psy_data wire: Proposal 214B, Certificate 204B, EndCapForwardHeader 60B,
  ValidatorLeafPreimage 108B; codecs and hash preimages widened to u64
- psy_node_core qblob: tree-node/blob headers widened to 84 bytes; all
  merkle views and the genesis builder threaded through u64
- psy_node_common runtime: realm/coordinator edge handlers, processors,
  GUTA planner, and consensus domain separation use full u64 magic
- psy_node_cli startup: u64 constructor wiring
- docs: documented incompatible storage/wire cohort in the operations guide

Storage from prior binaries is incompatible; purge required on upgrade.

+++ psy_cli/psy_node_cli/src/node/realm_p2p.rs
@@ -348,7 +348,7 @@ pub fn genesis_validator_index_from_network_config(
 /// Construct a processor Realm network from local keys/listen and public membership.
 pub fn maybe_build_processor_network(
     config: &RealmProcessorStartConfig,
-    chain_id: u32,
+    chain_id: u64,
 ) -> anyhow::Result<OptionalRealmNetwork> {
     let identity = config
         .p2p_identity_key_path
@@ -377,7 +377,7 @@ pub fn maybe_build_processor_network(
 /// Construct an edge Realm network from local identity/listen and public membership.
 pub fn maybe_build_edge_network(
     config: &RealmEdgeStartConfig,
-    chain_id: u32,
+    chain_id: u64,
 ) -> anyhow::Result<(OptionalRealmNetwork, HashMap<u16, NodeId>, HashSet<NodeId>, RealmRotationConfig)> {
     let identity = config
         .p2p_identity_key_path
